+++ demos/assets/redis-theme.json
@@ -52,6 +52,13 @@
     "background_fill_secondary": "#163341",
     "background_fill_primary_dark": "#091A23",
     "background_fill_secondary_dark": "#163341",
+    "checkbox_label_background_fill": "#091A23",
+    "checkbox_label_text_color": "#FFFFFF",
+    "checkbox_label_text_color_selected": "#FFFFFF",
+    "color_accent_soft": "#163341",
+    "border_color_accent_subdued": "#FFFFFF",
+    "input_background_fill": "#091A23",
+    "body_text_color": "#FFFFFF",
     "button_primary_background_fill": "#DCFF1E",
     "button_primary_background_fill_dark": "#DCFF1E",
     "button_primary_text_color": "#091A23",
